launch system of actors

+++ solar-panel-emulator/src/main/scala/ua/ucu/edu/model/Location.scala
@@ -1,3 +1,3 @@
 package ua.ucu.edu.model
 
-case class Location(latitude: Long, longitude: Long)
+case class Location(latitude: Double, longitude: Double)
